[vignetting] solving bugs

+++ tofu/geom/_vignetting_tools.pxd
@@ -24,13 +24,13 @@ cdef  bint is_pt_in_tri(double[3] v0, double[3] v1,
                         double px, double py, double pz) nogil
 
 cdef  void earclipping_poly(double[:,::1] vignett,
-                            int* ltri,
+                            long* ltri,
                             int nvert) nogil
 
 cdef void triangulate_polys(double[:, :, ::1] vignett_poly,
                                    long* lnvert,
                                    int nvign,
-                                   int** ltri,
+                                   long** ltri,
                                    int num_threads=*) nogil
 
 cdef void vignetting_core(double[:, ::1] ray_orig,
@@ -38,7 +38,7 @@ cdef void vignetting_core(double[:, ::1] ray_orig,
                                  double[:, :, ::1] vignett,
                                  long* lnvert,
                                  double* lbounds,
-                                 int** ltri,
+                                 long** ltri,
                                  int nvign,
                                  int nlos,
                                  bint* goes_through,
